+++ GloryEngine/GloryEntityScenes/EntitySceneScenesModule.h
@@ -26,7 +26,7 @@ namespace Glory
 
 		static EntitySceneObject* CreateDeserializedObject(GScene* pScene, const std::string& name, UUID uuid);
 
-		virtual SceneObject* GetSceneObjectFromObjectID(uint32_t objectID) override;
+		virtual SceneObject* GetSceneObjectFromObjectID(uint64_t objectID) override;
 
 	private:
 		friend class EntitySceneObjectSerializer;
